requirements.txt
# ── Core async ────────────────────────────────────────────────────────────────
aiofiles>=23.2.1
aiohttp>=3.9.1

# ── Audio ─────────────────────────────────────────────────────────────────────
faster-whisper>=1.0.0
numpy>=1.26.2
scipy>=1.11.4
librosa>=0.10.1
soundfile>=0.12.1
webrtcvad-wheels>=2.0.14     # pre-built wheels, no Visual C++ needed on Windows

# ── Vision ────────────────────────────────────────────────────────────────────
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80

# ── LLM ───────────────────────────────────────────────────────────────────────
openai>=1.14.0
tiktoken>=0.6.0

# ── TTS ───────────────────────────────────────────────────────────────────────
edge-tts>=6.1.9

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
websockets>=12.0
python-multipart>=0.0.9
httpx>=0.27.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client>=0.20.0
opentelemetry-sdk>=1.23.0
opentelemetry-api>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp>=1.23.0
structlog>=24.1.0

# ── Config & utilities ────────────────────────────────────────────────────────
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
tenacity>=8.2.3
